+++ src/components/views/index.js
@@ -1,7 +1,7 @@
 import Balance from './balance'
 import CreateWallet from './create-wallet'
 import CreateWalletForm from './create-wallet/create-wallet-form'
-import CreateWalletBackupWarning from './create-wallet/create-wallet-backup-warning'
+import BackupWarning from './backup/backup-warning'
 import CreateWalletBackupMnemonic from './create-wallet/create-wallet-backup-mnemonic'
 import CreateWalletMnemonicConfirm from './create-wallet/create-wallet-mnemonic-confirm'
 import CreateWalletMnemonicFailed from './create-wallet/create-wallet-mnemonic-failed'
@@ -27,7 +27,7 @@ import WarpPortal from './warp-portal'
 export {
   Balance,
   CreateWallet,
-  CreateWalletBackupWarning,
+  BackupWarning,
   CreateWalletForm,
   CreateWalletBackupMnemonic,
   CreateWalletMnemonicConfirm,
